refactor the helper files
