+++ public/lib/model/format-helpers.ts
@@ -1,6 +1,8 @@
-import { specialFormats, STANDARD_ARTICLE_FORMAT_LABEL } from './special-formats'
+import { specialFormats } from './special-formats'
 import { ContentType } from './stub'
 
+const STANDARD_ARTICLE_FORMAT_LABEL = "Standard Article";
+const STANDARD_ARTICLE_FORMAT_SHORT_LABEL = "Article";
 
 const nonArticleFormats = {
     "liveblog": "Live blog",
@@ -11,9 +13,13 @@ const nonArticleFormats = {
     "atom": "Video/Atom"
 }
 
+/**
+ * Returns an object mapping ContentType to user facings labels, excluding any
+ * special formats that that behind a feature switch in the 'off' state.
+ */
 const provideFormats = (featureSwitches?: Record<string, boolean>): Partial<Record<ContentType, string>> => {
     const articleFormats: Record<string, string> = {
-        "article": "Article",
+        "article": STANDARD_ARTICLE_FORMAT_SHORT_LABEL,
     }
 
     specialFormats.forEach(format => {
@@ -52,7 +58,27 @@ const provideArticleFormatsForDropDown = (featureSwitches?: Record<string, boole
     return list.map(label => ({ name: label, value: label }))
 }
 
+/**
+ * returns "Standard Article" for normal articles, the label for special article formats
+ * or empty string for non-article content types
+ */
+const getArticleFormatLabel = (contentType: ContentType): string => {
+    const maybeMatchingFormat = specialFormats.find(format => format.value === contentType)
+    if (maybeMatchingFormat) {
+        return maybeMatchingFormat.label
+    }
+    if (contentType === 'article') {
+        return STANDARD_ARTICLE_FORMAT_LABEL
+    }
+    return ''
+}
+
+/**
+ * true if the value is the label of a special format or the "Standard Article" label
+ */
+const isFormatLabel = (value: string): boolean => {
+    return value === STANDARD_ARTICLE_FORMAT_LABEL || specialFormats.some(format => format.label === value)
+}
 
 
-
-export { provideFormats, provideArticleFormatsForDropDown }
+export { provideFormats, provideArticleFormatsForDropDown, getArticleFormatLabel, isFormatLabel }
